Implement socket.io disconnect functionalitiy
