+++ server/index.js
@@ -63,12 +63,36 @@ const server = app.listen(PORT, () =>
 const io = socketio(server)
 io.on('connection', socket => {
   console.log(`A socket connection to the server has been made: ${socket.id}`)
+
   socket.on('join', ({name, room}) => {
     const {user} = addUser({id: socket.id, name, room})
-    console.log('THIS IS USER', user)
+
     socket.emit('message', {
       text: `${user.name}, welcome to the ${user.room} room!`
     })
+
+    socket.broadcast
+      .to(user.room)
+      .emit('message', {text: `${user.name} has joined!`})
+  })
+
+  socket.on('sendMessage', message => {
+    const user = getUser(socket.id)
+    io.to(user.room).emit('message', {user: user.name, text: message})
+    io
+      .to(user.room)
+      .emit('roomData', {room: user.room, users: getUsersInRoom(user.room)})
+  })
+
+  socket.on('disconnect', () => {
+    console.log(`Connection ${socket.id} has left the building`)
+    const user = removeUser(socket.id)
+    console.log('I CAN REMOVE USERS')
+    if (user) {
+      io
+        .to(user.room)
+        .emit('message', {user: 'admin', text: `${user.name} has left!`})
+    }
   })
 })
